Show links for both XML and JSON output.
Only a single c.uri_for() call is made, because using { inc => ..., fmt => ... } will not necessary emit the 2 parameters in the same order for both URLs.
But if this template is used in cases where inc ends up being empty, the concat with "&fmt=..." will not work.
A fix would be to adjust uri_for() to sort query arguments for deterministic results.